+++ examples/stm32wl/src/bin/button.rs
@@ -5,8 +5,8 @@ use core::mem::MaybeUninit;
 
 use cortex_m_rt::entry;
 use defmt::*;
-use embassy_stm32::gpio::{Input, Level, Output, Pull, Speed};
 use embassy_stm32::SharedData;
+use embassy_stm32::gpio::{Input, Level, Output, Pull, Speed};
 use {defmt_rtt as _, panic_probe as _};
 
 #[unsafe(link_section = ".shared_data")]